+++ LibCarla/source/carla/streaming/detail/Dispatcher.h
@@ -41,26 +41,30 @@ namespace detail {
 
     void DeregisterSession(std::shared_ptr<Session> session);
 
-    token_type GetToken(stream_id_type sensor_id);
+    token_type GetToken(stream_id_type stream_id);
 
-    void EnableForROS(stream_id_type sensor_id) {
-      auto search = _stream_map.find(sensor_id);
+    void SetROS2TopicVisibilityDefaultEnabled(bool topic_visibility_default_enabled) {
+      _topic_visibility_default_enabled = topic_visibility_default_enabled;
+    }
+
+    void EnableForROS(stream_actor_id_type stream_actor_id) {
+      auto search = _stream_map.find(stream_actor_id.stream_id);
       if (search != _stream_map.end()) {
-        search->second->EnableForROS();
+        search->second->EnableForROS(stream_actor_id.actor_id);
       }
     }
 
-    void DisableForROS(stream_id_type sensor_id) {
-      auto search = _stream_map.find(sensor_id);
+    void DisableForROS(stream_actor_id_type stream_actor_id) {
+      auto search = _stream_map.find(stream_actor_id.stream_id);
       if (search != _stream_map.end()) {
-        search->second->DisableForROS();
+        search->second->DisableForROS(stream_actor_id.actor_id);
       }
     }
 
-    bool IsEnabledForROS(stream_id_type sensor_id) {
-      auto search = _stream_map.find(sensor_id);
+    bool IsEnabledForROS(stream_actor_id_type stream_actor_id) {
+      auto search = _stream_map.find(stream_actor_id.stream_id);
       if (search != _stream_map.end()) {
-        return search->second->IsEnabledForROS();
+        return search->second->IsEnabledForROS(stream_actor_id.actor_id);
       }
       return false;
     }
@@ -74,6 +78,8 @@ namespace detail {
     token_type _cached_token;
 
     StreamMap _stream_map;
+
+    bool _topic_visibility_default_enabled{false};
   };
 
 } // namespace detail